+++ Changelog.md
@@ -1,4 +1,6 @@
 Change log:
+* 2024-11-14 - U41 - 2 newly available from Reward track 12
+  * Blonde Corgi, Parrot of the Fiery Sands
 * 2024-11-06 - U42 - 1 new + 2 newly available
   * Wild Black Donkey
   * Injured Bat, Jerboa